feat: Implement overflow rule expansion with RuleType enum and volume-based thresholds
Model changes:
- Added OverflowRuleType enum (SpecificResource=0, TotalWarehouse=1)
- Added RuleType property to WarehouseOverflowRule with StringEnumConverter
- Changed TriggerThreshold from int to decimal
- Updated ReadOnlyWarehouseOverflowRule to expose RuleType and decimal threshold
- Updated all consumers (BackgroundProcessor, FormColonyV2, Desktop ColonyViewModel)

Validation:
- Created OverflowRuleValidator static class with rule-type-aware validation
- SpecificResource requires ResourceName and ResourcePurity
- TotalWarehouse skips resource field validation

Evaluation logic:
- Refactored CheckWarehouseOverflow to switch/dispatch on RuleType
- Extracted EvaluateSpecificResourceRule with volume-based computation
- Implemented EvaluateTotalWarehouseRule with aggregate warehouse volume
- Added ComputeWarehouseVolume and GetItemUnitVolume helpers
- Unknown RuleType logs warning and skips

UI (FormColonyV2 Overflow tab):
- Added colOverflowRuleType column to dgvOverflowRules
- Added cmbOverflowRuleType ComboBox with SpecificResource/TotalWarehouse
- RuleType selection enables/disables resource+purity inputs
- PopulateOverflowGrid shows RuleType column, empty resource for TotalWarehouse
- CmdAddOverflowRule_Click uses OverflowRuleValidator, handles both rule types

Tests (13 total, all passing):
- Property tests: serialization round-trip, validation for both rule types, volume computation, excess calculation
- Unit tests: backward compatibility deserialization, default values, unknown RuleType handling

Spec: .kiro/specs/overflow-rule-expansion/

Prompt: Run all tasks for this spec.

+++ OE2EmpireTracker.Desktop/ViewModels/ColonyViewModel.cs
@@ -91,7 +91,7 @@ public sealed partial class ColonyOverflowRowViewModel : ObservableObject
     private string _purity = string.Empty;
 
     [ObservableProperty]
-    private int _threshold;
+    private decimal _threshold;
 
     [ObservableProperty]
     private string _destination = string.Empty;
